--- tests/sdk/tool/test_mcp_tool_json_serialization.py
@@ -1,216 +1,0 @@
-"""Test MCP tool JSON serialization with DiscriminatedUnionMixin.
-
-Note: MCPTool serialization may be limited due to complex MCP objects
-(mcp_tool field contains mcp.types.Tool which may not be fully JSON serializable).
-These tests demonstrate the expected behavior and limitations.
-"""
-
-import json
-from typing import Annotated
-from unittest.mock import Mock
-
-import pytest
-from pydantic import BaseModel, ValidationError
-
-import mcp.types
-from openhands.sdk.mcp.client import MCPClient
-from openhands.sdk.mcp.tool import MCPTool
-from openhands.sdk.tool import Tool, ToolType
-from openhands.sdk.utils.discriminated_union import DiscriminatedUnionType
-
-
-def create_mock_mcp_tool(name: str = "test_tool") -> mcp.types.Tool:
-    """Create a mock MCP tool for testing."""
-    return mcp.types.Tool(
-        name=name,
-        description=f"A test MCP tool named {name}",
-        inputSchema={
-            "type": "object",
-            "properties": {
-                "query": {"type": "string", "description": "Query parameter"}
-            },
-            "required": ["query"]
-        }
-    )
-
-
-def test_mcp_tool_json_serialization_limitations() -> None:
-    """Test that MCPTool JSON serialization has expected limitations due to complex objects."""
-    # Create mock MCP tool and client
-    mock_mcp_tool = create_mock_mcp_tool()
-    mock_client = Mock(spec=MCPClient)
-    
-    # Create MCPTool instance
-    mcp_tool = MCPTool.create(mock_mcp_tool, mock_client)
-    
-    # Attempting to serialize with mcp_tool field may fail due to complex objects
-    try:
-        tool_json = mcp_tool.model_dump_json()
-        # If serialization succeeds, test deserialization
-        deserialized_tool = MCPTool.model_validate_json(tool_json)
-        assert isinstance(deserialized_tool, MCPTool)
-        assert deserialized_tool.name == mcp_tool.name
-        assert deserialized_tool.description == mcp_tool.description
-    except Exception:
-        # Expected due to complex MCP objects that may not be JSON serializable
-        pass
-
-
-def test_mcp_tool_partial_json_serialization() -> None:
-    """Test MCPTool partial JSON serialization excluding complex fields."""
-    # Create mock MCP tool and client
-    mock_mcp_tool = create_mock_mcp_tool()
-    mock_client = Mock(spec=MCPClient)
-    
-    # Create MCPTool instance
-    mcp_tool = MCPTool.create(mock_mcp_tool, mock_client)
-    
-    # Serialize excluding potentially problematic fields
-    try:
-        partial_json = mcp_tool.model_dump_json(exclude={"mcp_tool", "executor"})
-        assert isinstance(partial_json, str)
-        
-        # Parse back to verify structure
-        parsed_data = json.loads(partial_json)
-        assert parsed_data["name"] == mcp_tool.name
-        assert parsed_data["description"] == mcp_tool.description
-    except Exception:
-        # May still fail due to action_type complexity
-        pass
-
-
-def test_mcp_tool_polymorphic_behavior() -> None:
-    """Test MCPTool polymorphic behavior using Tool base class."""
-    # Create mock MCP tool and client
-    mock_mcp_tool = create_mock_mcp_tool()
-    mock_client = Mock(spec=MCPClient)
-    
-    # Create MCPTool instance
-    mcp_tool = MCPTool.create(mock_mcp_tool, mock_client)
-    
-    # Should be instance of Tool
-    assert isinstance(mcp_tool, Tool)
-    assert isinstance(mcp_tool, MCPTool)
-    
-    # Check basic properties
-    assert mcp_tool.name == "test_tool"
-    assert "test MCP tool" in mcp_tool.description
-    assert hasattr(mcp_tool, 'mcp_tool')
-
-
-def test_mcp_tool_kind_field() -> None:
-    """Test that MCPTool kind field is correctly set."""
-    # Create mock MCP tool and client
-    mock_mcp_tool = create_mock_mcp_tool()
-    mock_client = Mock(spec=MCPClient)
-    
-    # Create MCPTool instance
-    mcp_tool = MCPTool.create(mock_mcp_tool, mock_client)
-    
-    # Check kind field
-    assert hasattr(mcp_tool, 'kind')
-    expected_kind = f"{mcp_tool.__class__.__module__}.{mcp_tool.__class__.__name__}"
-    assert mcp_tool.kind == expected_kind
-
-
-def test_mcp_tool_serializable_fields() -> None:
-    """Test that MCPTool core fields are accessible and serializable."""
-    # Create mock MCP tool and client
-    mock_mcp_tool = create_mock_mcp_tool("serializable_tool")
-    mock_client = Mock(spec=MCPClient)
-    
-    # Create MCPTool instance
-    mcp_tool = MCPTool.create(mock_mcp_tool, mock_client)
-    
-    # Extract core serializable fields
-    core_fields = {
-        "name": mcp_tool.name,
-        "description": mcp_tool.description,
-        "kind": mcp_tool.kind,
-        "annotations": mcp_tool.annotations,
-        "meta": mcp_tool.meta,
-    }
-    
-    # These should be JSON serializable
-    json_str = json.dumps(core_fields, default=str)  # default=str for any remaining complex objects
-    parsed_fields = json.loads(json_str)
-    
-    assert parsed_fields["name"] == "serializable_tool"
-    assert "serializable_tool" in parsed_fields["description"]
-    assert parsed_fields["kind"] == mcp_tool.kind
-
-
-def test_mcp_tool_in_container() -> None:
-    """Test MCPTool when used in a container with Tool type annotation."""
-    class ToolContainer(BaseModel):
-        tool: Tool
-    
-    # Create mock MCP tool and client
-    mock_mcp_tool = create_mock_mcp_tool("container_tool")
-    mock_client = Mock(spec=MCPClient)
-    
-    # Create MCPTool instance
-    mcp_tool = MCPTool.create(mock_mcp_tool, mock_client)
-    
-    # Should work as Tool in container
-    container = ToolContainer(tool=mcp_tool)
-    assert isinstance(container.tool, Tool)
-    assert isinstance(container.tool, MCPTool)
-    assert container.tool.name == "container_tool"
-
-
-def test_mcp_tool_fallback_behavior() -> None:
-    """Test MCPTool fallback behavior with manual data."""
-    # Create data that could represent an MCPTool
-    tool_data = {
-        "name": "fallback-tool",
-        "description": "A fallback test tool",
-        "action_type": "openhands.sdk.tool.schema.ActionBase",  # Use base class
-        "observation_type": "openhands.sdk.mcp.MCPToolObservation",
-        "kind": "openhands.sdk.mcp.tool.MCPTool",
-        "mcp_tool": {
-            "name": "fallback-tool",
-            "description": "A fallback test tool",
-            "inputSchema": {"type": "object", "properties": {}}
-        }
-    }
-    
-    try:
-        # This may work if MCP objects are properly serializable
-        deserialized_tool = Tool.model_validate(tool_data)
-        assert isinstance(deserialized_tool, Tool)
-        assert deserialized_tool.name == "fallback-tool"
-    except Exception:
-        # Expected due to complex MCP object structure
-        pass
-
-
-def test_mcp_tool_essential_properties() -> None:
-    """Test that MCPTool maintains essential properties after creation."""
-    # Create mock MCP tool with specific properties
-    mock_mcp_tool = mcp.types.Tool(
-        name="essential_tool",
-        description="Tool with essential properties",
-        inputSchema={
-            "type": "object",
-            "properties": {
-                "param1": {"type": "string"},
-                "param2": {"type": "integer"}
-            },
-            "required": ["param1"]
-        }
-    )
-    mock_client = Mock(spec=MCPClient)
-    
-    # Create MCPTool instance
-    mcp_tool = MCPTool.create(mock_mcp_tool, mock_client)
-    
-    # Verify essential properties are preserved
-    assert mcp_tool.name == "essential_tool"
-    assert mcp_tool.description == "Tool with essential properties"
-    assert mcp_tool.mcp_tool.name == "essential_tool"
-    assert mcp_tool.mcp_tool.inputSchema is not None
-    
-    # Verify action type was created correctly
-    assert mcp_tool.action_type is not None
-    assert hasattr(mcp_tool.action_type, 'to_mcp_arguments')
