+++ tests/sdk/tool/test_tool_json_serialization.py
@@ -1,34 +1,27 @@
 """Test tool JSON serialization with DiscriminatedUnionMixin."""
 
 import json
-from typing import Annotated
 
 from pydantic import BaseModel
 
 from openhands.sdk.tool import Tool, ToolType
 from openhands.sdk.tool.builtins import FinishTool, ThinkTool
-from openhands.sdk.utils.discriminated_union import DiscriminatedUnionType
 
 
-def test_tool_supports_polymorphic_json_serialization() -> None:
+def test_tool_serialization_deserialization() -> None:
     """Test that Tool supports polymorphic JSON serialization/deserialization."""
     # Use FinishTool which is a simple built-in tool
     tool = FinishTool
 
     # Serialize to JSON
     tool_json = tool.model_dump_json()
-    
+
     # Deserialize from JSON using the base class
     deserialized_tool = Tool.model_validate_json(tool_json)
 
-    # Should deserialize to the correct type with same core fields
+    # Should deserialize to the correct type with same serializable data
     assert isinstance(deserialized_tool, Tool)
-    assert deserialized_tool.name == tool.name
-    assert deserialized_tool.description == tool.description
-    assert deserialized_tool.action_type == tool.action_type
-    assert deserialized_tool.observation_type == tool.observation_type
-    assert deserialized_tool.annotations == tool.annotations
-    assert deserialized_tool.meta == tool.meta
+    assert tool.model_dump() == deserialized_tool.model_dump()
 
 
 def test_tool_supports_polymorphic_field_json_serialization() -> None:
@@ -43,16 +36,13 @@ def test_tool_supports_polymorphic_field_json_serialization() -> None:
 
     # Serialize to JSON
     container_json = container.model_dump_json()
-    
+
     # Deserialize from JSON
     deserialized_container = Container.model_validate_json(container_json)
 
-    # Should preserve the tool type with same core fields
+    # Should preserve the tool type with same serializable data
     assert isinstance(deserialized_container.tool, Tool)
-    assert deserialized_container.tool.name == tool.name
-    assert deserialized_container.tool.description == tool.description
-    assert deserialized_container.tool.action_type == tool.action_type
-    assert deserialized_container.tool.annotations == tool.annotations
+    assert tool.model_dump() == deserialized_container.tool.model_dump()
 
 
 def test_tool_supports_nested_polymorphic_json_serialization() -> None:
@@ -68,19 +58,16 @@ def test_tool_supports_nested_polymorphic_json_serialization() -> None:
 
     # Serialize to JSON
     container_json = container.model_dump_json()
-    
+
     # Deserialize from JSON
     deserialized_container = NestedContainer.model_validate_json(container_json)
 
-    # Should preserve all tool types and be exactly the same
+    # Should preserve all tool types with same serializable data
     assert len(deserialized_container.tools) == 2
     assert isinstance(deserialized_container.tools[0], Tool)
     assert isinstance(deserialized_container.tools[1], Tool)
-    assert deserialized_container.tools[0].name == tool1.name
-    assert deserialized_container.tools[1].name == tool2.name
-    assert deserialized_container.tools[0].action_type == tool1.action_type
-    assert deserialized_container.tools[1].action_type == tool2.action_type
-    # Container equality check removed due to computed fields
+    assert tool1.model_dump() == deserialized_container.tools[0].model_dump()
+    assert tool2.model_dump() == deserialized_container.tools[1].model_dump()
 
 
 def test_tool_model_validate_json_dict() -> None:
@@ -91,12 +78,13 @@ def test_tool_model_validate_json_dict() -> None:
     # Serialize to JSON, then parse to dict
     tool_json = tool.model_dump_json()
     tool_dict = json.loads(tool_json)
-    
+
     # Deserialize from dict
     deserialized_tool = Tool.model_validate(tool_dict)
 
-    # Should be exactly the same
-    # Tool equality check removed due to computed fields
+    # Should have same serializable data
+    assert isinstance(deserialized_tool, Tool)
+    assert tool.model_dump() == deserialized_tool.model_dump()
 
 
 def test_tool_fallback_behavior_json() -> None:
@@ -107,7 +95,7 @@ def test_tool_fallback_behavior_json() -> None:
         "description": "A test tool",
         "action_type": "openhands.sdk.tool.builtins.finish.FinishAction",
         "observation_type": None,
-        "kind": "UnknownToolType"
+        "kind": "UnknownToolType",
     }
     tool_json = json.dumps(tool_dict)
 
@@ -116,21 +104,6 @@ def test_tool_fallback_behavior_json() -> None:
     assert isinstance(deserialized_tool, Tool)
     assert deserialized_tool.name == "test-tool"
     assert deserialized_tool.description == "A test tool"
-
-
-def test_tool_preserves_pydantic_parameters_json() -> None:
-    """Test that Tool preserves Pydantic parameters through JSON serialization."""
-    # Create tool
-    tool = ThinkTool
-
-    # Serialize to JSON
-    tool_json = tool.model_dump_json()
-    
-    # Deserialize from JSON
-    deserialized_tool = Tool.model_validate_json(tool_json)
-
-    # Should preserve all fields exactly
-    # Tool equality check removed due to computed fields
 
 
 def test_tool_type_annotation_works_json() -> None:
@@ -146,33 +119,31 @@ def test_tool_type_annotation_works_json() -> None:
 
     # Serialize to JSON
     model_json = model.model_dump_json()
-    
+
     # Deserialize from JSON
     deserialized_model = TestModel.model_validate_json(model_json)
 
-    # Should work correctly
+    # Should work correctly with same serializable data
     assert isinstance(deserialized_model.tool, Tool)
-    assert deserialized_model.tool.name == tool.name
-    assert deserialized_model.tool.description == tool.description
-    # Model equality check removed due to computed fields
+    assert tool.model_dump() == deserialized_model.tool.model_dump()
 
 
 def test_tool_kind_field_json() -> None:
-    """Test that Tool kind field is correctly set and preserved through JSON serialization."""
+    """Test Tool kind field is correctly set and preserved through JSON."""
     # Create tool
     tool = FinishTool
 
     # Check kind field
-    assert hasattr(tool, 'kind')
+    assert hasattr(tool, "kind")
     expected_kind = f"{tool.__class__.__module__}.{tool.__class__.__name__}"
     assert tool.kind == expected_kind
 
     # Serialize to JSON
     tool_json = tool.model_dump_json()
-    
+
     # Deserialize from JSON
     deserialized_tool = Tool.model_validate_json(tool_json)
 
     # Should preserve kind field
-    assert hasattr(deserialized_tool, 'kind')
+    assert hasattr(deserialized_tool, "kind")
     assert deserialized_tool.kind == tool.kind
